+++ Kokoro/Features/PlayCardsFromAnywhere.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
+using Newtonsoft.Json.Converters;
 
 namespace Shockah.Kokoro;
 
@@ -50,19 +51,30 @@ partial class ApiImplementation
 
 			public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IPlayCardsFromAnywhereAction MakeAction(IEnumerable<(int CardId, Card? FallbackCard)> cards, int amount = 1)
 				=> new APlayRandomCardsFromAnywhere { Amount = amount }.SetCards(cards);
+
+			public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction? AsModifyAction(CardAction action)
+				=> action as IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction;
+
+			public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction MakeModifyAction(int cardId, CardAction action)
+				=> new AModifyCardAnywhere { CardId = cardId, Action = action };
 		}
 	}
 }
 
-internal sealed class PlaySpecificCardFromAnywhereManager
+internal static class PlaySpecificCardFromAnywhereManager
 {
 	internal static (Card Card, CardDestination OriginalDestination, int OriginalIndex)? CardBeingHackinglyPlayed;
+	internal static Card? CardBeingModified;
 	
 	internal static void Setup(IHarmony harmony)
 	{
 		harmony.Patch(
 			original: AccessTools.DeclaredMethod(typeof(Combat), nameof(Combat.TryPlayCard)),
 			transpiler: new HarmonyMethod(AccessTools.DeclaredMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(Combat_TryPlayCard_Transpiler)), priority: Priority.High)
+		);
+		harmony.Patch(
+			original: AccessTools.DeclaredMethod(typeof(Combat), nameof(Combat.RenderCards)),
+			postfix: new HarmonyMethod(AccessTools.DeclaredMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(Combat_RenderCards_Postfix_Last)), priority: Priority.Last)
 		);
 	}
 	
@@ -132,6 +144,19 @@ internal sealed class PlaySpecificCardFromAnywhereManager
 	{
 		if (card == CardBeingHackinglyPlayed?.Card)
 			state.RemoveCardFromWhereverItIs(card.uuid);
+	}
+
+	private static void Combat_RenderCards_Postfix_Last(G g)
+	{
+		if (CardBeingModified is not { } card)
+			return;
+
+		var cardUiKey = card.UIKey();
+		if (g.boxes.Any(b => b.key == cardUiKey))
+			return;
+		
+		card.Render(g);
+		CardBeingModified = null;
 	}
 }
 
@@ -307,6 +332,310 @@ public sealed class APlayRandomCardsFromAnywhere : CardAction, IKokoroApi.IV2.IP
 	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IPlayCardsFromAnywhereAction SetShowTheCardIfNotInHand(bool value)
 	{
 		ShowTheCardIfNotInHand = value;
+		return this;
+	}
+}
+
+public sealed class AModifyCardAnywhere : CardAction, IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction
+{
+	[JsonConverter(typeof(StringEnumConverter))]
+	private enum AnimationPhase
+	{
+		NotStarted,
+		ShowCard,
+		PrePause,
+		Modification,
+		PostPause,
+		HideCard,
+		Ended,
+	}
+	
+	public required int CardId { get; set; }
+	public required CardAction Action { get; set; }
+	public bool MoveIfInHand { get; set; }
+	public double PrePauseTime { get; set; } = TIMER_DEFAULT * 0.5;
+	public double PostPauseTime { get; set; } = TIMER_DEFAULT * 1.5;
+	public bool FlipOnBegin { get; set; } = true;
+	public bool FlipOnEnd { get; set; }
+	public bool FlopOnBegin { get; set; }
+	public bool FlopOnEnd { get; set; }
+
+	[JsonProperty]
+	private AnimationPhase Phase = AnimationPhase.NotStarted;
+
+	[JsonProperty]
+	private Vec OriginalCardTargetPosition;
+
+	[JsonProperty]
+	private Vec LastCardPosition;
+
+	[JsonProperty]
+	private double OriginalCardDrawAnimation;
+
+	[JsonProperty]
+	private double LastCardDrawAnimation;
+
+	[JsonProperty]
+	private double WaitingTime;
+	
+	[JsonIgnore]
+	public CardAction AsCardAction
+		=> this;
+
+	public override List<Tooltip> GetTooltips(State s)
+		=> Action.GetTooltips(s);
+
+	public override void Begin(G g, State s, Combat c)
+	{
+		base.Begin(g, s, c);
+		Phase = AnimationPhase.NotStarted;
+		
+		if (s.FindCard(CardId) is not { } card)
+		{
+			timer = 0;
+			return;
+		}
+
+		OriginalCardTargetPosition = card.targetPos;
+		OriginalCardDrawAnimation = card.drawAnim;
+		LastCardDrawAnimation = card.drawAnim;
+		SwitchToNextPhase(g, s, c, card);
+		PlaySpecificCardFromAnywhereManager.CardBeingModified = card;
+	}
+
+	public override void Update(G g, State s, Combat c)
+	{
+		if (Phase >= AnimationPhase.Ended || s.FindCard(CardId) is not { } card)
+		{
+			timer = 0;
+			PlaySpecificCardFromAnywhereManager.CardBeingModified = null;
+			return;
+		}
+
+		card.isForeground = true;
+		card.drawAnim = Mutil.MoveTowards(LastCardDrawAnimation, Phase is >= AnimationPhase.ShowCard and < AnimationPhase.HideCard ? 1 : OriginalCardDrawAnimation, g.dt * 10.0);
+		LastCardDrawAnimation = card.drawAnim;
+		PlaySpecificCardFromAnywhereManager.CardBeingModified = card;
+		
+		switch (Phase)
+		{
+			case AnimationPhase.NotStarted:
+				break;
+			case AnimationPhase.ShowCard:
+				HandleShowCardPhase(g, s, c, card);
+				break;
+			case AnimationPhase.PrePause:
+				HandlePrePausePhase(g, s, c, card);
+				break;
+			case AnimationPhase.Modification:
+				HandleModificationPhase(g, s, c, card);
+				break;
+			case AnimationPhase.PostPause:
+				HandlePostPausePhase(g, s, c, card);
+				break;
+			case AnimationPhase.HideCard:
+				HandleHideCardPhase(g, s, c, card);
+				break;
+			default:
+				throw new ArgumentOutOfRangeException();
+		}
+	}
+
+	private void SwitchToNextPhase(G g, State s, Combat c, Card card)
+	{
+		if (Phase == AnimationPhase.Ended)
+			return;
+		Phase++;
+		timer = 1;
+		LastCardPosition = card.pos;
+
+		switch (Phase)
+		{
+			case AnimationPhase.NotStarted:
+				break;
+			case AnimationPhase.ShowCard:
+				if (!MoveIfInHand && c.hand.Contains(card))
+					SwitchToNextPhase(g, s, c, card);
+				else
+					card.waitBeforeMoving = 0;
+				
+				break;
+			case AnimationPhase.PrePause:
+				if ((!MoveIfInHand && c.hand.Contains(card)) || PrePauseTime <= 0)
+					SwitchToNextPhase(g, s, c, card);
+				else
+					WaitingTime = PrePauseTime;
+				
+				break;
+			case AnimationPhase.Modification:
+				Action.selectedCard = card;
+				Action.Begin(g, s, c);
+				
+				if (FlipOnBegin)
+					card.flipAnim = 1;
+				if (FlopOnBegin)
+					card.flopAnim = 1;
+				
+				break;
+			case AnimationPhase.PostPause:
+				if ((!MoveIfInHand && c.hand.Contains(card)) || (PostPauseTime <= 0 && !FlipOnEnd && !FlopOnEnd))
+				{
+					SwitchToNextPhase(g, s, c, card);
+				}
+				else
+				{
+					WaitingTime = PostPauseTime;
+					if (FlipOnEnd)
+						card.flipAnim = 1;
+					if (FlopOnEnd)
+						card.flopAnim = 1;
+				}
+				
+				break;
+			case AnimationPhase.HideCard:
+				if (!MoveIfInHand && c.hand.Contains(card))
+				{
+					SwitchToNextPhase(g, s, c, card);
+				}
+				else
+				{
+					card.drawAnim = 1;
+					card.waitBeforeMoving = 0;
+				}
+				
+				break;
+			case AnimationPhase.Ended:
+				timer = 0;
+				PlaySpecificCardFromAnywhereManager.CardBeingModified = null;
+				break;
+			default:
+				throw new ArgumentOutOfRangeException();
+		}
+	}
+
+	private void HandleShowCardPhase(G g, State s, Combat c, Card card)
+	{
+		card.pos = LastCardPosition;
+		
+		UpdateCardPosition(g, card);
+		if (Math.Abs((card.targetPos - card.pos).len()) <= 1)
+			SwitchToNextPhase(g, s, c, card);
+	}
+
+	private void HandlePrePausePhase(G g, State s, Combat c, Card card)
+	{
+		card.pos = LastCardPosition;
+		card.drawAnim = 1;
+		card.waitBeforeMoving = card.flipAnim > 0 || card.flopAnim > 0 ? 0 : 1;
+		LastCardDrawAnimation = 1;
+		
+		WaitingTime -= g.dt;
+		if (WaitingTime <= 0)
+			SwitchToNextPhase(g, s, c, card);
+	}
+
+	private void HandleModificationPhase(G g, State s, Combat c, Card card)
+	{
+		card.pos = LastCardPosition;
+		card.drawAnim = 1;
+		card.waitBeforeMoving = card.flipAnim > 0 || card.flopAnim > 0 ? 0 : 1;
+		LastCardDrawAnimation = 1;
+
+		if (Action.timer > 0)
+		{
+			Action.selectedCard = card;
+			Action.Update(g, s, c);
+		}
+
+		if (card.flipAnim != 0 || card.flopAnim != 0)
+			return;
+		
+		SwitchToNextPhase(g, s, c, card);
+	}
+
+	private void HandlePostPausePhase(G g, State s, Combat c, Card card)
+	{
+		card.drawAnim = 1;
+		card.waitBeforeMoving = card.flipAnim > 0 || card.flopAnim > 0 ? 0 : 1;
+		LastCardDrawAnimation = 1;
+		
+		if (card.flipAnim != 0 || card.flopAnim != 0)
+		{
+			card.pos = LastCardPosition;
+			return;
+		}
+		
+		WaitingTime -= g.dt;
+		if (WaitingTime <= 0)
+			SwitchToNextPhase(g, s, c, card);
+	}
+
+	private void HandleHideCardPhase(G g, State s, Combat c, Card card)
+	{
+		UpdateCardPosition(g, card, OriginalCardTargetPosition);
+		
+		if (Math.Abs((card.targetPos - card.pos).len()) <= 1)
+			SwitchToNextPhase(g, s, c, card);
+	}
+
+	private void UpdateCardPosition(G g, Card card, Vec? targetPos = null)
+	{
+		card.targetPos = targetPos ?? new(Combat.cardCenter.x - 30, Combat.cardCenter.y - 120);
+		card.pos = Mutil.LerpDeltaSnap(LastCardPosition, card.targetPos, 12.0, g.dt);
+		LastCardPosition = card.pos;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetCardId(int value)
+	{
+		CardId = value;
+		return this;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetAction(CardAction value)
+	{
+		Action = value;
+		return this;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetMoveIfInHand(bool value)
+	{
+		MoveIfInHand = value;
+		return this;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetPrePauseTime(double value)
+	{
+		PrePauseTime = value;
+		return this;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetPostPauseTime(double value)
+	{
+		PostPauseTime = value;
+		return this;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetFlipOnBegin(bool value)
+	{
+		FlipOnBegin = value;
+		return this;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetFlipOnEnd(bool value)
+	{
+		FlipOnEnd = value;
+		return this;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetFlopOnBegin(bool value)
+	{
+		FlopOnBegin = value;
+		return this;
+	}
+
+	public IKokoroApi.IV2.IPlayCardsFromAnywhereApi.IModifyCardAnywhereAction SetFlopOnEnd(bool value)
+	{
+		FlopOnEnd = value;
 		return this;
 	}
 }